+++ packages/framework-integration-tests/integration/provider-unaware/end-to-end/read-models.integration.ts
@@ -1721,6 +1721,289 @@ describe('Read models end-to-end tests', () => {
         ])
       })
     })
+
+    context('query with fragments', () => {
+      const mockCartId: string = random.uuid()
+      const mockProductId: string = random.uuid()
+      const mockQuantity: number = random.number({ min: 1 })
+
+      beforeEach(async () => {
+        // provisioning a cart
+        await client.mutate({
+          variables: {
+            cartId: mockCartId,
+            productId: mockProductId,
+            quantity: mockQuantity,
+          },
+          mutation: gql`
+            mutation ChangeCartItem($cartId: ID!, $productId: ID!, $quantity: Float!) {
+              ChangeCartItem(input: { cartId: $cartId, productId: $productId, quantity: $quantity })
+            }
+          `,
+        })
+      })
+
+      it('should retrieve expected cart', async () => {
+        const fragment = gql`
+          fragment cartItemDetails on CartItem {
+            productId
+            quantity
+          }
+        `
+        const queryResult = await waitForIt(
+          () => {
+            return client.query({
+              variables: {
+                cartId: mockCartId,
+              },
+              query: gql`
+                query CartReadModel($cartId: ID!) {
+                  CartReadModel(id: $cartId) {
+                    id
+                    cartItems {
+                      ...cartItemDetails
+                    }
+                  }
+                }
+                ${fragment}
+              `,
+            })
+          },
+          (result) => result?.data?.CartReadModel != null
+        )
+
+        const cartData = queryResult.data.CartReadModel
+
+        expect(cartData.id).to.be.equal(mockCartId)
+        expect(cartData.cartItems).to.have.length(1)
+        expect(cartData.cartItems[0]).to.deep.equal({
+          __typename: 'CartItem',
+          productId: mockProductId,
+          quantity: mockQuantity,
+        })
+      })
+
+      it('should retrieve list of items', async () => {
+        const fragment = gql`
+          fragment cart on CartReadModel {
+            id
+            cartItems {
+              productId
+              quantity
+            }
+          }
+        `
+        const limit = 1
+        let cursor: Record<'id', string> | undefined = undefined
+
+        for (let i = 0; i < limit; i++) {
+          const queryResult = await waitForIt(
+            () => {
+              return client.query({
+                variables: {
+                  filterBy: { id: { eq: mockCartId } },
+                },
+                query: gql`
+                  query ListCartReadModels($filterBy: ListCartReadModelFilter) {
+                    ListCartReadModels(filter: $filterBy) {
+                      items {
+                        ...cart
+                      }
+                      cursor
+                    }
+                  }
+                  ${fragment}
+                `,
+              })
+            },
+            (result) => result?.data?.ListCartReadModels?.items.length === 1
+          )
+
+          const currentPageCartData = queryResult.data.ListCartReadModels.items
+
+          cursor = queryResult.data.ListCartReadModels.cursor
+
+          if (cursor) {
+            if (process.env.TESTED_PROVIDER === 'AZURE' || process.env.TESTED_PROVIDER === 'LOCAL') {
+              expect(cursor.id).to.equal((i + 1).toString())
+            } else {
+              expect(cursor.id).to.equal(currentPageCartData[0].id)
+            }
+          }
+          expect(currentPageCartData).to.be.an('array')
+          expect(currentPageCartData.length).to.equal(1)
+          expect(cursor).to.not.be.undefined
+          expect(currentPageCartData[0].id).to.be.equal(mockCartId)
+          expect(currentPageCartData[0].cartItems).to.have.length(1)
+          expect(currentPageCartData[0].cartItems[0].productId).to.equal(mockProductId)
+        }
+      })
+    })
+
+    context('projecting calculated fields', () => {
+      const mockCartId: string = random.uuid()
+      const mockProductId: string = random.uuid()
+      const mockQuantity: number = random.number({ min: 1 })
+      const mockAddress = {
+        firstName: random.word(),
+        lastName: random.word(),
+        country: random.word(),
+        state: random.word(),
+        postalCode: random.word(),
+        address: random.word(),
+      }
+
+      beforeEach(async () => {
+        // provisioning a cart
+        await client.mutate({
+          variables: {
+            cartId: mockCartId,
+            productId: mockProductId,
+            quantity: mockQuantity,
+          },
+          mutation: gql`
+            mutation ChangeCartItem($cartId: ID!, $productId: ID!, $quantity: Float!) {
+              ChangeCartItem(input: { cartId: $cartId, productId: $productId, quantity: $quantity })
+            }
+          `,
+        })
+
+        await waitForIt(
+          () => {
+            return client.query({
+              variables: {
+                cartId: mockCartId,
+              },
+              query: gql`
+                query CartReadModel($cartId: ID!) {
+                  CartReadModel(id: $cartId) {
+                    id
+                  }
+                }
+              `,
+            })
+          },
+          (result) => result?.data?.CartReadModel != null
+        )
+
+        // update shipping address
+        await client.mutate({
+          variables: {
+            cartId: mockCartId,
+            address: mockAddress,
+          },
+          mutation: gql`
+            mutation UpdateShippingAddress($cartId: ID!, $address: AddressInput!) {
+              UpdateShippingAddress(input: { cartId: $cartId, address: $address })
+            }
+          `,
+        })
+
+        await waitForIt(
+          () => {
+            return client.query({
+              variables: {
+                cartId: mockCartId,
+              },
+              query: gql`
+                query CartReadModel($cartId: ID!) {
+                  CartReadModel(id: $cartId) {
+                    id
+                    shippingAddress {
+                      firstName
+                    }
+                  }
+                }
+              `,
+            })
+          },
+          (result) =>
+            result?.data?.CartReadModel != null &&
+            result?.data?.CartReadModel?.shippingAddress?.firstName === mockAddress.firstName
+        )
+      })
+
+      it('should correctly fetch calculated fields via GraphQL query', async () => {
+        const queryResult = await waitForIt(
+          () => {
+            return client.query({
+              variables: {
+                filter: {
+                  id: { eq: mockCartId },
+                },
+              },
+              query: gql`
+                query CartReadModels($filter: CartReadModelFilter) {
+                  CartReadModels(filter: $filter) {
+                    id
+                    myAddress {
+                      firstName
+                      lastName
+                      country
+                      state
+                      postalCode
+                      address
+                    }
+                  }
+                }
+              `,
+            })
+          },
+          (result) => result?.data?.CartReadModels?.length >= 1
+        )
+
+        const cartData = queryResult.data.CartReadModels
+
+        expect(cartData).to.be.an('array')
+        expect(cartData.length).to.equal(1)
+        expect(cartData[0].id).to.be.equal(mockCartId)
+        expect(cartData[0].myAddress).to.deep.equal({
+          ...mockAddress,
+          __typename: 'Address',
+        })
+      })
+
+      it('should correctly fetch calculated fields via code', async () => {
+        const queryResult = await waitForIt(
+          () => {
+            return client.mutate({
+              variables: {
+                cartId: mockCartId,
+                paginatedVersion: true,
+              },
+              mutation: gql`
+                mutation CartMyAddress($cartId: ID!, $paginatedVersion: Boolean!) {
+                  CartMyAddress(input: { cartId: $cartId, paginatedVersion: $paginatedVersion })
+                }
+              `,
+            })
+          },
+          (result) => result?.data?.CartMyAddress != null
+        )
+
+        const cartMyAddress = queryResult.data.CartMyAddress
+
+        expect(cartMyAddress).to.deep.equal({
+          items: [
+            {
+              id: mockCartId,
+              myAddress: {
+                firstName: mockAddress.firstName,
+                lastName: mockAddress.lastName,
+                country: mockAddress.country,
+                state: mockAddress.state,
+                postalCode: mockAddress.postalCode,
+                address: mockAddress.address,
+              },
+            },
+          ],
+          count: 1,
+          cursor: {
+            id: '1',
+          },
+        })
+      })
+    })
   })
 
   describe('projections', () => {